Comments addressed in relation to refactoring bson readers
