+++ app/repositories/onlinetesting/OnlineTestRepository.scala
@@ -28,12 +28,14 @@ import org.joda.time.{ DateTime, DateTimeZone }
 import play.api.Logger
 import reactivemongo.bson.{ BSONDocument, _ }
 import repositories._
+import repositories.application.BaseBSONReader
 import uk.gov.hmrc.mongo.ReactiveRepository
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-trait OnlineTestRepository extends RandomSelection with BSONHelpers with CommonBSONDocuments with OnlineTestCommonBSONDocuments {
+trait OnlineTestRepository extends RandomSelection with BSONHelpers with CommonBSONDocuments with OnlineTestCommonBSONDocuments
+  with BaseBSONReader {
   this: ReactiveRepository[_, _] =>
 
   val thisApplicationStatus: ApplicationStatus
