+++ it/repositories/ApplicationRepositorySpec.scala
@@ -29,7 +29,7 @@ import model.report.AdjustmentReportItem
 import org.joda.time.DateTime
 import reactivemongo.bson.BSONDocument
 import reactivemongo.json.ImplicitBSONHandlers
-import repositories.application.{ GeneralApplicationMongoRepository, GeneralApplicationRepoBSONToModelHelper, TestDataMongoRepository }
+import repositories.application.{ GeneralApplicationMongoRepository, GeneralApplicationRepoBSONReader, TestDataMongoRepository }
 import repositories.assistancedetails.AssistanceDetailsMongoRepository
 import services.GBTimeZoneService
 import testkit.MongoRepositorySpec
@@ -44,7 +44,7 @@ class ApplicationRepositorySpec extends MongoRepositorySpec {
 
   val collectionName = "application"
 
-  def applicationRepo = new GeneralApplicationMongoRepository(GBTimeZoneService, cubiksGatewayConfig, GeneralApplicationRepoBSONToModelHelper)
+  def applicationRepo = new GeneralApplicationMongoRepository(GBTimeZoneService, cubiksGatewayConfig)
 
   def assistanceRepo = new AssistanceDetailsMongoRepository()
