+++ it/repositories/personaldetails/PersonalDetailsRepositorySpec.scala
@@ -6,7 +6,7 @@ import model.persisted.PersonalDetailsExamples._
 import org.joda.time.DateTime
 import reactivemongo.bson.BSONDocument
 import reactivemongo.json.ImplicitBSONHandlers
-import repositories.application.{GeneralApplicationMongoRepository, GeneralApplicationRepoBSONToModelHelper}
+import repositories.application.{GeneralApplicationMongoRepository}
 import services.GBTimeZoneService
 import config.MicroserviceAppConfig._
 import testkit.MongoRepositorySpec
@@ -17,7 +17,7 @@ class PersonalDetailsRepositorySpec extends MongoRepositorySpec {
   override val collectionName = "application"
 
   def repository = new PersonalDetailsMongoRepository
-  def appRepository = new GeneralApplicationMongoRepository(GBTimeZoneService, cubiksGatewayConfig, GeneralApplicationRepoBSONToModelHelper)
+  def appRepository = new GeneralApplicationMongoRepository(GBTimeZoneService, cubiksGatewayConfig)
 
   "update candidate" should {
     "modify the details and find the personal details successfully" in {
